# storage/spanlatch: create spanlatch.Manager using immutable btrees