--- a/pkg/storage/spanlatch/interval_btree.go
+++ b/pkg/storage/spanlatch/interval_btree.go
@@ -53,9 +53,9 @@ func cmp(a, b *latch) int {
 	if c != 0 {
 		return c
 	}
-	if a.id() < b.id() {
+	if a.id < b.id {
 		return -1
-	} else if a.id() > b.id() {
+	} else if a.id > b.id {
 		return 1
 	} else {
 		return 0