--- a/pkg/storage/spanlatch/interval_btree_test.go
+++ b/pkg/storage/spanlatch/interval_btree_test.go
@@ -12,7 +12,7 @@
 // implied. See the License for the specific language governing
 // permissions and limitations under the License.
 
-package cmdq
+package spanlatch
 
 import (
 	"fmt"