--- a/pkg/storage/spanlatch/signal_test.go
+++ b/pkg/storage/spanlatch/signal_test.go
@@ -12,7 +12,7 @@
 // implied. See the License for the specific language governing
 // permissions and limitations under the License.
 
-package cmdq
+package spanlatch
 
 import (
 	"sync"