@nvanbenschoten 2018-10-30
There's still plenty of room to optimize the btree implementation that we use here. Until this point we've closely follower `github.com/google/btree`, but there are some good ideas that we could pull from [pebble's btree](https://github.com/petermattis/pebble/blob/master/internal/btree/btree.go). For instance, we could inline item and node pointer arrays in nodes to avoid unnecessary indirection. We could also specialize the implementation to this use to avoid interfaces. Short of this, we could at least cache the interface's `Range()` value to avoid repeatedly calling the method when searching.