@nvanbenschoten 2018-11-22
I've update this PR to use the interval btree type and the signal type added in #32165, #32251, #32164 to create a new `spanlatch.Manager`. The commits here don't replace the `CommandQueue` yet, but that's the immediate follow-up to this PR.